+++ Rag_5040/BLM.lua
@@ -8,7 +8,7 @@ local ninSJMaxMP = nil -- The Max MP you have when /nin in your idle set
 local whmSJMaxMP = nil -- The Max MP you have when /whm in your idle set
 local rdmSJMaxMP = nil -- The Max MP you have when /rdm in your idle set
 
-local nukeExtraThreshold = 850 -- The minimum MP for which NukeExtra StoneskinExtra, and PhalanxExtra set will be used instead of regular sets (to allow additional nukes using max mp sets)
+local extraThreshold = 850 -- The minimum MP for which NukeExtra StoneskinExtra, and PhalanxExtra set will be used instead of regular sets (to allow additional nukes using max mp sets or specific defensive buffing above idle max mp thresholds)
 
 -- Comment out the equipment within these sets if you do not have them or do not wish to use them
 local warlocks_mantle = { -- Don't add 2% to fastCastValue for this as it is SJ dependant
@@ -540,24 +540,16 @@ profile.HandleWeaponskill = function()
 end
 
 profile.OnLoad = function()
-    gcinclude.SetAlias(T{'extra'})
-    gcdisplay.CreateToggle('Extra', false)
     gcmage.Load(3.00)
     profile.SetMacroBook()
 end
 
 profile.OnUnload = function()
     gcmage.Unload()
-    gcinclude.ClearAlias(T{'extra'})
 end
 
 profile.HandleCommand = function(args)
-    if (args[1] == 'extra') then
-        gcdisplay.AdvanceToggle('Extra')
-        gcinclude.Message('Extra', gcdisplay.GetToggle('Extra'))
-    else
-        gcmage.DoCommands(args, sets)
-    end
+    gcmage.DoCommands(args, sets)
 
     if (args[1] == 'horizonmode') then
         profile.HandleDefault()
@@ -598,13 +590,13 @@ end
 local ElementalDebuffs = T{ 'Burn','Rasp','Drown','Choke','Frost','Shock' }
 
 profile.HandleMidcast = function()
-    gcmage.DoMidcast(sets, ninSJMaxMP, whmSJMaxMP, nukeExtraThreshold, rdmSJMaxMP, nil)
+    gcmage.DoMidcast(sets, ninSJMaxMP, whmSJMaxMP, extraThreshold, rdmSJMaxMP, nil)
 
     local player = gData.GetPlayer()
     local action = gData.GetAction()
     if (republic_circlet.Head) then
         if (action.Skill == 'Elemental Magic' and gcdisplay.GetCycle('Mode') == 'Potency') then
-            if (gcdisplay.GetToggle('Extra') and player.MP >= nukeExtraThreshold) then
+            if (gcdisplay.GetToggle('Extra') and player.MP >= extraThreshold) then
                 do return end
             end
             if (not ElementalDebuffs:contains(action.Name)) then