+++ common/changelog.md
@@ -649,4 +649,5 @@
 
 ## v3.0.6
 - RNG - Changes special_ammo to a table to allow multiple different weapon types to be supported within a single lua
+- RDM / WHM - Implements a gated /extra and /mb command
 - Fixes debuff or cancel functionality targeting